create outputs like DSMEngine but keep inversion
Mimic DSMEngine startup as far as connecting outputs from the XML, but keep
the inversion of control where the HotFilm class reads the data then publishes
the samples directly to the outputs.  This allows a server socket output to
provide samples in real-time (although still technically float samples and not
"raw") and also write to a disk archive.  Add -n arg.