refactor sample ids, set timetag from pps counter
give each channel and the stats for each channel their own SPS ID, to make it
easier to differentiate channel values without needing to know how many
channels were recorded.  use changes in the pps count to assign the time tag
to the samples, where the timetag is the time of the first measurement in the
sample.  add min and max to the stats for each channel, and also add pps and
backlog stats.