reset time adjustment on each block
the point was to adjust the sample times in each block so they were contiguous
and exactly 500us apart, even if the labjack itself was recording samples a
little faster or slower than exactly 2 KHz.  each block is supposed to start
at the actual time of the first sample, since that is the closest time to the
exact time the block starts, but I forgot to reset the time adjustment on each
new block.