add read_scans() generator method, fix bugs
rather than keep an entire DataFrame in memory and keep appending to it,
convert to a generator method which yields each scan, once the minimum time
period of scans is reached.  use this to write text files scan by scan.  write
more tests, find more bugs: get_scan() now returns the pending scan when it
hits the end of the data.