insert minutes into final output filename
write to a temporary file first, then rename to a final filename which inserts
the number of minutes in the file into the filename before the suffix.  this
also means all files without the temporary suffix are known to be complete.
the default minimum block length is now 1 minute, since the point is to write
out all continuous blocks as files and then select blocks by duration.