reorganize directories to clean up top level
the nidas c++ hotfilm program is now in src, the python tests and test data
files are under `tests`, a pyproject.toml file is used to set paths for
pytest, tests still pass.  the python modules have moved to `hotfilm`, but not
sure yet if they should become a package.