fix times parsed by xarray by adding tod
xarray seems to only parse the date out of the time units in isfs netcdf
files, presumably because of the space, and the only workaround I could find
was to parse the time of day separately and replace the time coordinate.  also
get_variable() now always returns the reshaped array, since that is more
useful, and the reshaped array includes the original attributes.