replace pandas.DataFrame with xarray.Dataset, #3
datetime objects were replaced with numpy datetime64 and timedelta64, which do
not have as many format and conversion methods, so pandas is still used in
several places to convert back and forth where needed.  numpy datetime64 and
timedelta64 are also not timezone aware, so explicit utc was removed all over
the place and utc is assumed instead.

skip_scan() was especially troubling since any() works on a dataset but
broadcasts always return a DataArray so the result was always true.

remove unused time_selected() method.

tried to update type hints everywhere.

the Dataset attributes and time coordinate replacement should probably be
deferred until right before writing to netcdf, but not done yet.