add sample_rate to hotfilm voltage attributes
check for sample_rate and break up blocks when the rate changes, then store
the rate in the netcdf as an attribute of the hotfilm voltage variables.  also
fix integer attributes to use np.int32 type, so they are encoded in netcdf as
an int type instead of long long (LL).