add rsquared to hotfilm calibrations
also update test data to include uvw since the hotfilm wind speed dataset
now requires v to include it in the netcdf outpute; add tests for r-squared
calculation, and test writing a calibrated hotfilm wind speed dataset
netcdf file, but so far the test only checks for existence and not for
content.

Two R^2 values are calculated, one for the linear fit (where the terms of
the fit are exponentials of the measured variables) and one where the
residuals are calculated in the wind speed domain, between the sonic wind
speeds and the calibrated hotfilm wind speeds.  Only the latter is
currently included in the netcdf output.