use isel() not sel() to extract time intervals
for reasons I was not able to figure out, sel() sometimes raises
exceptions when indexing the dataset to separate time intervals, so
replace it with isel().

it is not necessary to concatenate each scan into the dataset, and that
seems much slower, so revert back to collecting scans in a list then
concatenating them in one call.

do not write a new file if all it would contain are the remainders of
one second of data leftover from the previous file.